requirements.txt
fastapi
uvicorn
rembg
pillow
